move to a new section
