+++ src/banner/styling.js
